Fix label in game
