+++ books.json
@@ -1,3 +1,1 @@
-{ "publisher": "la france parle", "cover_state": "bad", "id": "288", "genre": "book_genre_name-1", "author_first_name": "guy", "author_last_name": "cars", "source": "france", "label_title": "corruptrice", "label_color": "red", "publish_date": "1996/01/24", "archive": "false"} 
-{ "publisher": "sand dunes", "cover_state": "good", "id": "560", "genre": "book_genre_name-1", "author_first_name": "samm", "author_last_name": "mia", "source": "lusaka", "label_title": "le desert", "label_color": "yellow", "publish_date": "2000/01/31", "archive": "false"} 
-{ "publisher": "orange", "cover_state": "good", "id": "586", "genre": "book_genre_name-1", "author_first_name": "amadou", "author_last_name": "djaili", "source": "douala", "label_title": "patience", "label_color": "white", "publish_date": "2003/01/23", "archive": "false"} 
+{ "publisher": "francais", "cover_state": "bad", "id": "368", "genre": "book_genre_name-1", "author_first_name": "guy", "author_last_name": "cars", "source": "france", "label_title": "corruptrice", "label_color": "red", "publish_date": "1996/06/21", "archive": "false"}
